requirements.txt
numpy>=1.26.4
streamlit>=1.37.1
pandas>=2.2.2
plotly-express>=0.4.1
openpyxl>=3.1.2
matplotlib>=3.8
scikit-learn>=1.5.1
oauth2client>=4.1.3
gspread>=6.2.1
spotipy>=2.25.1
python-dotenv>=1.1.1
xlsxwriter>=3.2.5
vaderSentiment==3.3.2


